making gh_action.sh to run locally

+++ platform/linux/gh_action.sh
@@ -7,30 +7,36 @@ sudo lxd init --minimal
 
 rm -rf docs/SampleCode/.git docs/SampleCode/.gitignore
 
-cp -v Webtemplate/webtemplate.zip platform/resources/webtemplate.zip
-
-tar -xvzf Native/CoronaNative.tar.gz CoronaEnterprise/Corona/android/resource/android-template.zip CoronaEnterprise/Corona/android/lib/gradle/Corona.aar
-
-
-
-SNAPDIR=platform/linux/snapcraft/snap
-sed -i "s|2100|${YEAR:=2100}|" $SNAPDIR/snapcraft.yaml
-if [[ "${BUILD_NUMBER:=9999}" == "9999" ]]
+if [ -e Webtemplate/webtemplate.zip ]
 then
-    sed -i "s|stable|devel|"  $SNAPDIR/snapcraft.yaml
-    if [ "$GITHUB_SHA" ]
-    then
-         sed -i "s|9999|9999.${GITHUB_SHA::7}|" $SNAPDIR/snapcraft.yaml
-    fi
-else
-    sed -i "s|9999|$BUILD_NUMBER|" $SNAPDIR/snapcraft.yaml
+    cp -v Webtemplate/webtemplate.zip platform/resources/webtemplate.zip
 fi
 
-mkdir -p output
+if [ -e Native/CoronaNative.tar.gz ]
+then
+    tar -xvzf Native/CoronaNative.tar.gz CoronaEnterprise/Corona/android/resource/android-template.zip CoronaEnterprise/Corona/android/lib/gradle/Corona.aar
+fi
 
-rm -rf ./*.snap
-cp -Rv $SNAPDIR ./
+
+rm -rf ./*.snap ./snap
+cp -Rv platform/linux/snapcraft/snap ./
+
+sed -i "s|2100|${YEAR:=2100}|" snap/snapcraft.yaml
+if [[ "${BUILD_NUMBER:=9999}" == "9999" ]]
+then
+    sed -i "s|stable|devel|"  snap/snapcraft.yaml
+    if [ "$GITHUB_SHA" ]
+    then
+         sed -i "s|9999|9999.${GITHUB_SHA::7}|" snap/snapcraft.yaml
+    fi
+else
+    sed -i "s|9999|$BUILD_NUMBER|" snap/snapcraft.yaml
+fi
+
+
 sudo snapcraft --use-lxd
+
+mkdir -p output
 cp -v ./*.snap output/
 
 FS2D="$(mktemp -d)"
